# docs/swagger.yaml
definitions:
...
  models.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      load_id:
        type: integer
      rating:
        description: 1-5 stars
        type: integer
      review_type:
        description: CARRIER_TO_SHIPPER, SHIPPER_TO_CARRIER
        type: string
      reviewee_id:
        type: integer
      reviewer_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Vehicle:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      has_liftgate:
        type: boolean
      has_straps:
        type: boolean
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      inspection_expiry:
        type: string
      insurance_expiry:
        type: string
      is_active:
        type: boolean
      is_food_grade:
        type: boolean
      is_hazmat_certified:
        type: boolean
      is_refrigerated:
        type: boolean
      license_plate:
        type: string
      load_capacity_kg:
        type: number
      load_capacity_m3:
        type: number
      make:
        type: string
      max_height:
        type: number
      max_length:
        type: number
      max_width:
        type: number
      model:
        type: string
      registration_expiry:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      vehicle_type:
        description: FLATBED, REEFER, DRY_VAN, TANKER, BOX_TRUCK
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
info:
  contact: {}
  title: Triplink API
  version: "1.0"
paths:
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Create a new review for a user after a completed load
      parameters:
      - description: Review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Review'
      summary: Create a review
      tags:
      - reviews
  /reviews/{id}:
    delete:
      description: Delete a review (only by the reviewer)
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete review
      tags:
      - reviews
    get:
      description: Get a specific review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      summary: Get review by ID
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Update an existing review (only by the reviewer)
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      summary: Update review
      tags:
      - reviews
...
  /users/{user_id}/rating-summary:
    get:
      description: Get rating summary and statistics for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get user rating summary
      tags:
      - reviews
  /users/{user_id}/reviews:
    get:
      description: Get all reviews for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      summary: Get reviews for a user
      tags:
      - reviews
  /users/{user_id}/reviews-given:
    get:
      description: Get all reviews written by a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      summary: Get reviews by a user
      tags:
      - reviews
  /users/{user_id}/vehicles:
    get:
      description: Get all vehicles belonging to a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vehicle'
            type: array
      summary: Get all vehicles for a user
      tags:
      - vehicles
  /vehicles:
    post:
      consumes:
      - application/json
      description: Create a new vehicle for a carrier
      parameters:
      - description: Vehicle data
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/models.Vehicle'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vehicle'
      summary: Create a new vehicle
      tags:
      - vehicles
  /vehicles/{id}:
    delete:
      description: Soft delete a vehicle (set is_active to false)
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete vehicle
      tags:
      - vehicles
    get:
      description: Get a specific vehicle by its ID
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vehicle'
      summary: Get vehicle by ID
      tags:
      - vehicles
    put:
      consumes:
      - application/json
      description: Update an existing vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated vehicle data
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/models.Vehicle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vehicle'
      summary: Update vehicle
      tags:
      - vehicles
  /vehicles/search:
    get:
      description: Search for available vehicles by type, capacity, and location
      parameters:
      - description: Vehicle type
        in: query
        name: vehicle_type
        type: string
      - description: Minimum weight capacity
        in: query
        name: min_capacity_kg
        type: number
      - description: Minimum volume capacity
        in: query
        name: min_capacity_m3
        type: number
      - description: Requires liftgate
        in: query
        name: has_liftgate
        type: boolean
      - description: Requires refrigeration
        in: query
        name: is_refrigerated
        type: boolean
      - description: Requires hazmat certification
        in: query
        name: is_hazmat_certified
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vehicle'
            type: array
      summary: Search vehicles
      tags:
      - vehicles
swagger: "2.0"